// frontend/src/backend.d.ts
export type Time = bigint;
export interface Spending {
    id: string;
    date: Time;
    description: string;
    category: SpendingCategory;
    amount: bigint;
}
export interface PatientSession {
    id: string;
    checkIn: Time;
    duration: bigint;
    visitNumber: bigint;
    progressPercentage: bigint;
    patientId: string;
    date: Time;
    treatmentPackageId: string;
    checkOut?: Time;
}